backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-adk==1.4.2
fal-client==0.4.1
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6